added the picture for the banner. it's wrong but i have no clue how to make it right, since i haven't been taught how to do anything of the sort.
'fixed' the positioning of some headers, but i know
that they are not 'properly' done. But hey, not like
ive been taught how to properly do it.

added profile pictures to the news posts, with 'correct'
styling for that part of those news cards.
but i know i will be told i did it wrong, or that
it is incorrect somehow. But again, no guidance =
inadequate results.

i don't feel like a web developer.
this course has been nothing but pain, anguish, and
has triggered panic attacks as well as resurfaced
my depression which i had almost dealt with.